Refactor nix flake, based on overriding nixpkgs
Now that umu-launcher-unwrapped is being maintained in nixpkgs, it can
be used as a starting point for the nix package here. We just need to
override a few things, such as the `src`, and handle any changes that
have not yet made it into a release build.

Ideally, we would also reuse the nixpkgs umu-launcher wrapper package.
However, 1) it is currently not fully overridable and 2) we have a
slightly different package-args API that needs to be maintained.

The opportunity has been taken to clean up that package-args API:
- Renamed `truststore` to `withTruststore` with a warning.
- `cbor2` wasn't working, warn when it is used.
- Added `withDeltaUpdates`

Breaking changes:
- The umu-launcher package in the overlay is no longer the unwrapped
  package (previously in `umu-launcher.nix`).
- The derivation attrs for various packages are re-written from scratch,
  likely breaking users' custom overrides.

Non-breaking changes:
- The umu & umu-run packages have effectively been renamed umu-launcher.
- The flake outputs for all systems supported by the nixpkgs package.
- The package version now defaults to the flake's git revision.